+++ client/vnctl/lib/vnctl/cli/route.rb
@@ -17,7 +17,7 @@ module Vnctl::Cli
     add_modify_shared_options
     option :ingress, :type => :boolean, :desc => "Flag to determine if this is an ingress route."
     option :egress, :type => :boolean, :desc => "Flag to determine if this is an egress route."
-    set_required_options [:ipv4_address, :network_uuid, :route_link_uuid]
+    set_required_options [:network_uuid, :route_link_uuid]
     define_add
 
     add_modify_shared_options