+++ vnet/lib/vnet/node_api/segment.rb
@@ -15,6 +15,8 @@ module Vnet::NodeApi
 
         filter = { segment_id: model.id }
 
+        # 0009_topology
+        TopologySegment.dispatch_deleted_where(filter, model.deleted_at)
         # 0011_assoc_interface
         InterfaceSegment.dispatch_deleted_where(filter, model.deleted_at)
       end